requirements.txt
python-telegram-bot==20.7
gspread==5.12.4
google-auth==2.40.2
oauth2client==4.0.0
gunicorn==22.0.0
Flask==3.0.0
Werkzeug==3.1.3
httpx==0.25.2